+++ src/tests/hla_plugin_test/UnicodeStringTest.cpp
@@ -17,7 +17,8 @@ BOOST_FIXTURE_TEST_CASE( unicode_string_serializes_size_and_content, Serializati
 {
     const std::string message = "message";
     const UnicodeString unicode( message );
-    ::hla::Deserializer deserializer = Serialize( unicode, sizeof( uint32 ) + message.size() * sizeof( uint16 ) );
+    ::hla::Deserializer deserializer = Serialize( unicode,
+        static_cast< unsigned int >( sizeof( uint32 ) + message.size() * sizeof( uint16 ) ));
     BOOST_CHECK_EQUAL( message.size(), Read< uint32 >( deserializer ) );
     BOOST_CHECK_EQUAL( wchar_t( 'm' ), Read< uint16 >( deserializer ) );
 }
@@ -26,7 +27,8 @@ BOOST_FIXTURE_TEST_CASE( unicode_string_deserializes, SerializationFixture )
 {
     const std::string message = "message";
     const UnicodeString serializedUnicode( message );
-    ::hla::Deserializer deserializer = Serialize( serializedUnicode, sizeof( uint32 ) + message.size() * sizeof( uint16 ) );
+    ::hla::Deserializer deserializer = Serialize( serializedUnicode,
+        static_cast< unsigned int >( sizeof( uint32 ) + message.size() * sizeof( uint16 ) ));
     UnicodeString deserializedUnicode;
     deserializedUnicode.Deserialize( deserializer );
     BOOST_CHECK_EQUAL( message, deserializedUnicode.str() );
